security 적용 후  에러나는 현상 Fixes #47
