+++ src/main/java/dev/ohhoonim/component/package-info.java
@@ -1,4 +1,5 @@
-@org.springframework.modulith.ApplicationModule(type=org.springframework.modulith.ApplicationModule.Type.OPEN)
+@org.springframework.modulith.ApplicationModule
+(type=org.springframework.modulith.ApplicationModule.Type.OPEN)
 package dev.ohhoonim.component;
 
 /**
