+++ src/main/java/dev/ohhoonim/configuration/package-info.java
@@ -1,0 +1,4 @@
+@org.springframework.modulith.ApplicationModule(allowedDependencies = {
+    "component::signJwt"
+})
+package dev.ohhoonim.configuration;
