+++ src/main/resources/application.properties
@@ -27,3 +27,4 @@ spring.modulith.moments.granularity=hours
 # for secutiry test
 logging.level.org.springframework.security=trace
 
+# spring.aop.proxy-target-class=true
